move media query to view
